// frontend/src/types/Auth.ts
export interface User {
    _id: string;
    name: string;
    email: string;
  }
  
  export interface AuthResponse {
    _id: string;
    name: string;
    email: string;
    token: string;
  }
  
  export interface LoginCredentials {
    email: string;
    password: string;
  }
  
  export interface RegisterCredentials {
    name: string;
    email: string;
    password: string;
  }